C Language Tutorial for Beginners
Welcome to the C Language Tutorial for Beginners! This repository contains a series of lessons designed to teach you the fundamentals of the C programming language.

Overview
This tutorial is aimed at beginners who have little or no programming experience. It starts with the basics of setting up a development environment and writing your first "Hello, World!" program. From there, we will cover the following topics:

Variables and data types
Operators and expressions
Control structures (if/else, switch/case, loops)
Functions
Arrays
Pointers
Structs and unions
File I/O
Each lesson builds on the previous one, so it's recommended that you go through them in order. We'll also provide plenty of examples and exercises to help you practice and reinforce your understanding.

Prerequisites
Before you get started, you'll need:

A computer running Windows, macOS, or Linux
A text editor or integrated development environment (IDE) such as Visual Studio Code, Sublime Text, or Code::Blocks
A C compiler installed on your computer, such as GCC or Clang
If you're not sure how to set up your development environment, don't worry! We'll provide step-by-step instructions in the first lesson.

How to Use This Tutorial
This tutorial is organized into a series of lessons, each contained in its own directory. Each lesson includes a README file that provides an overview of the material covered, as well as example code and exercises to complete.

To get started, navigate to the lesson-01 directory and follow the instructions in the README file. Once you've completed the exercises for a given lesson, you can move on to the next one.

Contributing
We welcome contributions to this tutorial! If you find an error or have a suggestion for improvement, please open an issue or submit a pull request.

License
Atul kashyap